+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All the changes, additions, bugfixes, etc. made to this project.
 ## 2022-01-10
 ### Added
 - Base filtering by `complete` status
+- Fully implement a show `complete` toggle with it's corresponding `settingsReducer` and group of options and a `<TodoSettings />` component
 
 ### Modified
 - Changed the way the tasks to render filtering is coded for a clearer and cleaner code